+++ fitsio/hdu/__init__.py
@@ -1,0 +1,9 @@
+from .base import (  # noqa
+    ANY_HDU, BINARY_TBL, ASCII_TBL, IMAGE_HDU, _hdu_type_map)
+from .image import ImageHDU  # noqa
+from .table import (  # noqa
+    TableHDU,
+    AsciiTableHDU,
+    _table_npy2fits_form,
+    _npy2fits,
+)
